refactor: modularize ffmpeg and fs helpers; document code structure
- Extract ffmpeg-related functions to tools/ffmpeg_utils.py (audio extract, burn-in, font detection, ffprobe).
- Extract filesystem helpers to tools/fs_utils.py (ensure_dirs, find_videos, extensions).
- Keep CLI stable by delegating legacy functions in tools/subtitle_gen.py to new modules.
- Add tools/__init__.py to mark package.
- Update README with a brief Code Structure section.
