+++ tools/__init__.py
@@ -1,1 +1,2 @@
+"""Helper modules for the subtitle generator CLI."""
